[MCUX-85803] Fix MSG issues
Address static analysis warnings (Coverity/MSG) across multiple files:

- Add missing NULL pointer checks and return statements
- Replace unsafe strtol() with shell_strtol() and add error handling
- Add explicit type casts to fix implicit conversion warnings
- Add input parameter range validation
- Fix potential integer overflow/underflow issues

Files modified:
- lib/asrc_sw/srCvtFrm.c
- source/impl/ethermind/host/bt_pal_crypto.c
- source/impl/ethermind/host/bt_pal_hci_core.c
- source/services/wu.c
- source/shell/shell_bredr.c
- source/shell/shell_test_mode.c

+++ source/impl/ethermind/host/bt_pal_crypto.c
@@ -599,7 +599,7 @@ int bt_aes_128_cmac_update(bt_aes_128_cmac_state_t *state, const uint8_t * data,
 		len -= sizeof(state->data) - state->len;
 		input = &data[sizeof(state->data) - state->len];
 		state->len = sizeof(state->data);
-		n = (uint8_t) ((len) / sizeof(state->data)); /* n is number of rounds */
+		n = (uint32_t) ((len) / sizeof(state->data)); /* n is number of rounds */
 		if (sizeof(state->data) == len)
 		{
 			n = 0;
